Adds scheme_factory
- Scheme class is now replaced by the Pydantic model
- Add a factory method
